LayoutTests/http/tests/inspector/network/intercept-request-subresource-with-error-expected.txt

CONSOLE MESSAGE: Access denied
Test request interception with error for the subresource load.


== Running test suite: Network.interceptRequestWithError
-- Running test case: Network.interceptRequest.Subresource with general error
Triggering load...
FAILURE TEXT: Request intercepted

-- Running test case: Network.interceptRequest.Subresource with access error
Triggering load...
FAILURE TEXT: Access denied

-- Running test case: Network.interceptRequest.Subresource with canceled error
Triggering load...
FAILURE TEXT: Request canceled

-- Running test case: Network.interceptRequest.Subresource with timeout error
Triggering load...
FAILURE TEXT: Request timed out
